requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pillow==11.3.0
ptl==0.3.0
requests==2.32.5
urllib3==2.5.0
